# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: RF-R3x0 REST RFID Hardware Interface
  version: 1.0.0
  description: This is a REST based interface to interact with the RF-R3x0 RFID Hardware Interface via http using JSON.<br>
    It is designed as a low level reader interface in accordance with LLRP 1.1 (Low Level Reader Protocol) by GS1.
  contact: 
    email: support@peraMIC.io
servers:
  - url: https://{mica-host}/Apps/rest/rf
    description: MICA
    variables:
      mica-host:
        default: mica-host
        description: This value depends on the MICA on which the app is installed
security:
  - basic_auth: []
components:
  securitySchemes:
    basic_auth:
      type: http
      scheme: basic
  schemas:
...
    DeviceCapabilities:
      type: object
      description: Represents the device specific capabilities of the reader
      properties:
        manufacturer:
            type: integer
            format: int32
            description: The manufacturer name of the reader as IANA private enterprise number (PEN)
        model:
            type: integer
            format: int32
            description: The model name of the reader as unsigned integer
        firmware:
          type: string
          description: The hardware firmware version of the reader as UTF-8 string
        numberOfAntennas:
          type: integer
          format: int16
          description: The maximum number of antennas provided by the reader
        maxReceiveSensitivity:
          type: integer
          format: int16
          description: The maximum receive sensitity supported by the reader in dBm
        receiveSensitivityTable:
          $ref: "#/components/schemas/ReceiveSensitivityTable"
        antennaReceiveSensitivityRangeTable:
          $ref: "#/components/schemas/AntennaReceiveSensitivityRangeTable"
...
    RegulatoryCapabilities:
      type: object
      description: Represents the RF regulation specific capabilities. This include regulatory standard, frequency band information, power levels supporzted, frquencies supported and any air protocol specific values.
      properties:
        countryCode:
          type: integer
          format: int16
          description: The numeric code of the country as specified in ISO 3166, while value 0 means unspecified.
        communicationStandard:
          type: integer
          format: int16
          description: The communication standard as specified by LLRP specification 1.1 chapter 10.2.4
        transmitPowerTable:
          $ref: "#/components/schemas/TransmitPowerTable"
        hopping:
          type: boolean
          description: True if frequency hopping is mandatory for the current communication standard, false otherwise.
        freqHopTable:
          $ref: "#/components/schemas/FreqHopTable"
        fixedFreqTable:
          $ref: "#/components/schemas/FixedFreqTable"
...
    AntennaProperties:
      type: object
      description: Represents a single antenna's properties.
      properties:
        id:
          type: integer
          format: int16
          description: The antenna ID
        gain:
          type: integer
          format: int16
          description: >-
            Not supported by RF-R3x0.
            The gain of the antenna in dBI*100
        connected:
          type: boolean
          description: True if antenna port is connected or activated via AntennaConfiguration, false otherwise
    
    #configuration 
...
  parameters:
...
    antennaID:
      name: antennaID
      in: path
      description: The ID of an specific antenna
      schema:
        type: number
      required: true
...
paths:
...
  /capabilities/device:
    get:
      tags: 
        - capabilities
      summary: Gets the device capabilities
      operationId: getDeviceCapabilities
      responses:
        '200':
          description: Device capabilities
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/DeviceCapabilities"
        '500':
          description: Unexpected error occured
  /capabilities/regulatory:
    get:
      tags: 
        - capabilities
      summary: Gets the regulatory capabilities
      operationId: getRegulatoryCapabilities
      responses:
        '200':
          description: List of regulatory capabilities
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/RegulatoryCapabilities"
        '500':
          description: Unexpected error occured
  /capabilities/antennaProperties:
    get:
      tags: 
        - capabilities
      summary: Gets the antenna properties for all antennas
      operationId: getAntennaPropertiesList
      responses:
        '200':
          description: List of antenna properties
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AntennaProperties"
        '500':
          description: Unexpected error occured
  /capabilities/antennaProperties/{antennaID}:
    get:
      tags: 
        - capabilities
      summary: Gets the antenna properties of the specified antenna
      operationId: getAntennaProperties
      parameters: 
        - $ref: '#/components/parameters/antennaID'
      responses:
        '200':
          description: Antenna properties
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/AntennaProperties" 
        '500':
          description: Unexpected error occured
  /configuration:
    get:
      tags: 
        - configuration
      summary: Gets the complete configuration of the reader in one document
      operationId: getConfiguration
      responses:
        '200':
          description: Complete device configuration
          content:
            'application/octet-stream':
              schema:
                $ref: "#/components/schemas/DeviceConfiguration"
        '500':
          description: Unexpected error occured
    post:
      tags: 
        - configuration
      summary: Sets the comple configuration at once
      operationId: setConfiguration
      requestBody:
        content:
          'application/octet-stream':
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: OK
        '400':
          description: Invalid configuration
        '500':
          description: Unexpected error occured
...
  /configuration/antennas:
    get:
      tags: 
        - configuration
      summary: Gets the antenna configurations for all antennas
      operationId: getAntennaConfigurationList
      responses:
        '200':
          description: List of antenna configurations
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AntennaConfiguration"
        '500':
          description: Unexpected error occured
  /configuration/antennas/{antennaID}:
    get:
      tags: 
        - configuration
      summary: Gets the antenna configuration of the specified antenna
      operationId: getAntennaConfiguration
      parameters: 
        - $ref: '#/components/parameters/antennaID'
      responses:
        '200':
          description: Antenna configurations
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/AntennaConfiguration"
        '400':
          description: Invalid antenna configuration
        '500':
          description: Unexpected error occured
    put:
      tags: 
        - configuration
      summary: Sets the antenna configuration of the specified antenna
      operationId: setAntennaConfiguration
      parameters: 
        - $ref: '#/components/parameters/antennaID'
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/AntennaConfiguration'
      responses:
        '200':
          description: OK
        '500':
          description: Unexpected error occured
  /configuration/inventory:
    get:
      tags: 
        - configuration
      summary: Gets the complete inventory settings of the reader
      operationId: getInventorySettings
      responses:
        '200':
          description: Inventory settings
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/InventorySettings'
        '500':
          description: Unexpected error occured
  /configuration/inventory/selectionMasks:
    get:
      tags: 
        - configuration
      summary: Gets the selection mask list of the reader
      operationId: getSelectionMasks
      responses:
        '200':
          description: List of selection masks
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SelectionMask'
        '500':
          description: Unexpected error occured
    put:
      tags: 
        - configuration
      summary: Sets the selection mask list of the reader
      operationId: setSelectionMasks
      requestBody:
        content:
          'application/json':
            schema:
              type: array
              items:
                $ref: '#/components/schemas/SelectionMask'
      responses:
        '200':
          description: OK
        '400':
          description: Invalid selection mask
        '500':
          description: Unexpected error occured
  /configuration/inventory/rssiFilter:
    get:
      tags: 
        - configuration
      summary: Gets the RSSI filter of the reader
      operationId: getRssiFilter
      responses:
        '200':
          description: Return object
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/RssiFilter'
        '500':
          description: Unexpected error occured
    put:
      tags: 
        - configuration
      summary: Sets the RSSI filter of the reader
      operationId: setRssiFilter
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/RssiFilter'
      responses:
        '200':
          description: OK
        '400':
          description: Invalid rssi value or maxRssi smaller then minRssi
        '500':
          description: Unexpected error occured
  /configuration/inventory/singulationControl:
    get:
      tags: 
        - configuration
      summary: Gets the singulation control settings of th reader
      operationId: getSingulationControl
      responses:
        '200':
          description: The singulation control
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/SingulationControl'
        '500':
          description: Unexpected error occured
    put:
      tags: 
        - configuration
      summary: Sets the singulation control settings of the reader
      operationId: setSingulationControl
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/SingulationControl'
      responses:
        '200':
          description: OK
        '400':
          description: Invalud singulation control
        '500':
          description: Unexpected error occured
  /configuration/inventory/tagSmoothing:
    get:
      tags: 
        - configuration
      summary: Gets the tag smoothing settings of the reader
      operationId: getTagSmoothingSettings
      responses:
        '200':
          description: Return object
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/TagSmoothingSettings'
        '500':
          description: Unexpected error occured
    put:
      tags: 
        - configuration
      summary: Sets the tag smoothing settings of the reader
      operationId: setTagSmoothingSettings
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/TagSmoothingSettings'
      responses:
        '200':
          description: OK
        '400':
          description: Invalid tag smoothing configuration
        '500':
          description: Unexpected error occured
  /region:
    get:
      tags: 
        - configuration
      summary: Gets the configured region code of the reader
      operationId: getRegion
      responses:
        '200':
          description: Region code
          content:
            'text/plain':
              schema:
                type: string
        '500':
          description: Unexpected error occured
    put:
      tags: 
        - configuration
      summary: Sets the region of the reader
      operationId: setRegion
      requestBody:
        content:
          'text/plain':
            schema:
              type: string
      responses:
        '200':
          description: OK
        '400':
          description: Invalid region or region was already set
        '500':
          description: Unexpected error occured
  /region/supported:
    get:
      tags: 
        - configuration
      summary: Gets the supported regions of the reader
      operationId: getSupportedRegions
      responses:
        '200':
          description: Return object
          content:
            'application/json':
              schema:
                type: array
                items:
                  type: string
        '500':
          description: Unexpected error occured
  /tags:
    get:
      tags: 
        - operations
      summary: Gets all rfid tags that are currently within the antenna field based on the configuration of the reader
      description: This will perform an rfid inventory at the activated antennas based on the configured inventory seeting. The result is a list of captured rfid tags that matches the specified filter conditions and optionally one specific antenna if the antenna query parameter is given.
      operationId: getTags
      parameters:
        - $ref: '#/components/parameters/antenna'
      responses:
        '200':
          description: List of tag data
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TagData"
                example:
                  - tagDataId: 0
                    epc: "300833B2DDD9014000000000"
                    pc: 13312
                    crc: 0
                    xpc: 0
                    antennaID: 1
                    rssi: -49
                    channel: 1
                    resultList: []
                  - tagDataId: 1
                    epc: "340D964A3120000000000002"
                    pc: 13312
                    crc: 0
                    xpc: 0
                    antennaID: 1
                    rssi: -49
                    channel: 1
                    resultList: []
        '400':
          description: Invalid Operation
        '500':
          description: Unexpected error occured
...